+++ tests/DefaultConfigContainsAppUrlTest.php
@@ -8,11 +8,12 @@ use Orchestra\Testbench\TestCase;
 
 class DefaultConfigContainsAppUrlTest extends TestCase
 {
-    protected function useDefaultStatefulConfiguration($app)
+    protected function getEnvironmentSetUp($app)
     {
+        putenv('APP_URL=https://www.example.com');
         $config = require __DIR__.'/../config/sanctum.php';
 
-        $app->config->set('sanctum.stateful', $config['stateful']);
+        $app['config']->set('sanctum.stateful', $config['stateful']);
     }
 
     public function test_default_config_contains_app_url()
@@ -24,12 +25,20 @@ class DefaultConfigContainsAppUrlTest extends TestCase
         $this->assertContains($app_host, $config['stateful']);
     }
 
-    /**
-     * @environment-setup useDefaultStatefulConfiguration
-     */
+    public function test_app_url_is_not_parsed_when_missing_from_env()
+    {
+        putenv('APP_URL');
+
+        $config = require __DIR__.'/../config/sanctum.php';
+
+        $this->assertNull(env('APP_URL'));
+        $this->assertNotContains('', $config['stateful']);
+    }
+
     public function test_request_from_app_url_is_stateful_with_default_config()
     {
         $request = Request::create('/');
+
         $request->headers->set('referer', env('APP_URL'));
 
         $this->assertTrue(EnsureFrontendRequestsAreStateful::fromFrontend($request));